+++ babel.config.js
@@ -21,6 +21,7 @@ module.exports = function(api) {
       ["@babel/plugin-proposal-decorators", { legacy: true }],
       ["@babel/plugin-proposal-class-properties", { loose: true }],
       ["relay", { artifactDirectory: "./src/__generated__" }],
+      ["import-graphql"]
     ],
     presets: ["module:metro-react-native-babel-preset", "@babel/preset-typescript"],
   }