+++ src/__generated__/ArtworkTombstoneStoryQuery.graphql.ts
@@ -1,0 +1,245 @@
+/* tslint:disable */
+
+import { ConcreteRequest } from "relay-runtime";
+import { ArtworkTombstone_artwork$ref } from "./ArtworkTombstone_artwork.graphql";
+export type ArtworkTombstoneStoryQueryVariables = {};
+export type ArtworkTombstoneStoryQueryResponse = {
+    readonly artwork: {
+        readonly " $fragmentRefs": ArtworkTombstone_artwork$ref;
+    } | null;
+};
+export type ArtworkTombstoneStoryQuery = {
+    readonly response: ArtworkTombstoneStoryQueryResponse;
+    readonly variables: ArtworkTombstoneStoryQueryVariables;
+};
+
+
+
+/*
+query ArtworkTombstoneStoryQuery {
+  artwork(id: "unused") {
+    ...ArtworkTombstone_artwork
+    id
+  }
+}
+
+fragment ArtworkTombstone_artwork on Artwork {
+  title
+  medium
+  date
+  cultural_maker
+  artists {
+    name
+    href
+    ...FollowArtistButton_artist
+    id
+  }
+  dimensions {
+    in
+    cm
+  }
+  edition_of
+  attribution_class {
+    short_description
+    id
+  }
+}
+
+fragment FollowArtistButton_artist on Artist {
+  gravityID
+  id
+  is_followed
+}
+*/
+
+const node: ConcreteRequest = (function(){
+var v0 = [
+  {
+    "kind": "Literal",
+    "name": "id",
+    "value": "unused"
+  }
+],
+v1 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "id",
+  "args": null,
+  "storageKey": null
+};
+return {
+  "kind": "Request",
+  "fragment": {
+    "kind": "Fragment",
+    "name": "ArtworkTombstoneStoryQuery",
+    "type": "Query",
+    "metadata": null,
+    "argumentDefinitions": [],
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "artwork",
+        "storageKey": "artwork(id:\"unused\")",
+        "args": (v0/*: any*/),
+        "concreteType": "Artwork",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "FragmentSpread",
+            "name": "ArtworkTombstone_artwork",
+            "args": null
+          }
+        ]
+      }
+    ]
+  },
+  "operation": {
+    "kind": "Operation",
+    "name": "ArtworkTombstoneStoryQuery",
+    "argumentDefinitions": [],
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "artwork",
+        "storageKey": "artwork(id:\"unused\")",
+        "args": (v0/*: any*/),
+        "concreteType": "Artwork",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "title",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "medium",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "date",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "cultural_maker",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "LinkedField",
+            "alias": null,
+            "name": "artists",
+            "storageKey": null,
+            "args": null,
+            "concreteType": "Artist",
+            "plural": true,
+            "selections": [
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "name",
+                "args": null,
+                "storageKey": null
+              },
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "href",
+                "args": null,
+                "storageKey": null
+              },
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "gravityID",
+                "args": null,
+                "storageKey": null
+              },
+              (v1/*: any*/),
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "is_followed",
+                "args": null,
+                "storageKey": null
+              }
+            ]
+          },
+          {
+            "kind": "LinkedField",
+            "alias": null,
+            "name": "dimensions",
+            "storageKey": null,
+            "args": null,
+            "concreteType": "dimensions",
+            "plural": false,
+            "selections": [
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "in",
+                "args": null,
+                "storageKey": null
+              },
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "cm",
+                "args": null,
+                "storageKey": null
+              }
+            ]
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "edition_of",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "LinkedField",
+            "alias": null,
+            "name": "attribution_class",
+            "storageKey": null,
+            "args": null,
+            "concreteType": "AttributionClass",
+            "plural": false,
+            "selections": [
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "short_description",
+                "args": null,
+                "storageKey": null
+              },
+              (v1/*: any*/)
+            ]
+          },
+          (v1/*: any*/)
+        ]
+      }
+    ]
+  },
+  "params": {
+    "operationKind": "query",
+    "name": "ArtworkTombstoneStoryQuery",
+    "id": "dc28e3077bde459644be71da863d85c1",
+    "text": null,
+    "metadata": {}
+  }
+};
+})();
+(node as any).hash = '9441dbaf2c30e7220cf0a6885e46df60';
+export default node;